Covert to typescript the unarchive slash commands files
