--- app/slashcommands-unarchiveroom/client/client.js
@@ -1,7 +1,0 @@
-import { slashCommands } from '../../utils';
-
-slashCommands.add('unarchive', null, {
-	description: 'Unarchive',
-	params: '#channel',
-	permission: 'unarchive-room',
-});
